Move apollo-client to app.js
